requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
africastalking==1.2.5
paho-mqtt==2.1.0
pymysql==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
email-validator==2.1.0
marshmallow==3.20.1
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.3